+++ components/inbox-mail/inboxMailLocalTranslations.json
@@ -1,14 +1,35 @@
 {
     "inboxMail": {
+        "header": "Inbox Mail",
+        "headerDescription": "Manage your inbox messages.",
+        "getSupport": "Get Support",
+        "newMail": "New Mail",
+        "noMails": "No inbox mails",
+        "noMailsInfo": "You have no mails in your inbox",
+        "reply": "Reply",
+        "organization": "Organization",
+        "orgSelectionPlaceholder": "Select organization",
+        "mailSelectionPlaceholder": "Select mail to see the details",
+        "kernAITeam": "Kern AI Team",
+        "searchPlaceholder": "Search by name or email...",
+        "deleteHeader": "Delete Mail",
+        "deleteConfirmation": "Are you sure you want to delete this inbox mail? This action cannot be undone.",
+        "supportRequest": "Support Request",
+        "highPriority": "High Priority",
+        "inProgress": "In Progress",
+        "resolved": "Resolved",
+        "newBadge": "new",
+        "unknownUser": "Unknown User",
         "modalTitle": "Send New Inbox Mail",
         "sendTo": "Send To",
         "subject": "Subject",
         "content": "Content",
-        "isImportant": "Mark as Important",
+        "isImportant": "Important",
         "isImportantInfo": "Use this option to highlight messages that require immediate attention.",
         "includeProjectInfo": "Reference project",
         "includeChatInfo": "Reference chat",
         "sendButton": "Send",
+        "deleteButton": "Delete",
         "cancelButton": "Cancel",
         "previous": "Previous",
         "next": "Next"
